+++ ebong/src/app/component/VoteProfileList/VoteProfileList.tsx
@@ -1,19 +1,19 @@
 "use client";
 
-import { useContext } from "react";
 import Image from "next/image";
-import { VoteProfileContext } from "@/provider/vote-profile-provider";
+import { useVoteProfileList } from "./VoteProfileListHooks";
 
 export const VoteProfileList = () => {
-  const { voteProfiles, vote } = useContext(VoteProfileContext);
+  const { data } = useVoteProfileList();
+  const voteProfiles = data?.vote_profiles;
 
-  const handleVote = async (_id: string) => {
-    try {
-      await vote(_id);
-    } catch (error) {
-      console.error("Error voting:", error);
-    }
-  };
+  // const handleVote = async (_id: string) => {
+  //   try {
+  //     await vote(_id);
+  //   } catch (error) {
+  //     console.error("Error voting:", error);
+  //   }
+  // };
   return (
     <>
       {voteProfiles?.map((profile) => {
@@ -24,7 +24,7 @@ export const VoteProfileList = () => {
             <h1>{name}</h1>
             <p>{introduction}</p>
             <p>{voteCount}</p>
-            <button onClick={() => handleVote(_id)}>투표하기</button>
+            {/* <button onClick={() => handleVote(_id)}>투표하기</button> */}
           </div>
         );
       })}
