+++ ebong/src/app/component/VoteProfileList/VoteProfileListHooks.ts
@@ -1,0 +1,9 @@
+import { useQuery } from "@tanstack/react-query";
+import { voteProfileClientApi } from "@/lib/client-api/vote-profiles";
+
+export const useVoteProfileList = () => {
+  return useQuery({
+    queryKey: ["voteProfiles"],
+    queryFn: voteProfileClientApi.getVoteProfiles
+  });
+};
