+++ ebong/src/app/vote/page.tsx
@@ -1,9 +1,22 @@
-import { VoteProfileList } from "../component/VoteProfileList";
+import { VoteProfileList } from "../component/VoteProfileList/VoteProfileList";
+import { voteProfileClientApi } from "@/lib/client-api/vote-profiles";
+import {
+  HydrationBoundary,
+  QueryClient,
+  dehydrate,
+} from "@tanstack/react-query";
 
-export default function VoteResultPage() {
+export default async function VoteResultPage() {
+  const queryClient = new QueryClient();
+
+  await queryClient.prefetchQuery({
+    queryKey: ["voteProfiles"],
+    queryFn: voteProfileClientApi.getVoteProfiles,
+  });
+
   return (
-    <>
+    <HydrationBoundary state={dehydrate(queryClient)}>
       <VoteProfileList />
-    </>
+    </HydrationBoundary>
   );
 }
